+++ base/src/test/kotlin/proguard/analysis/cpa/ExecutingInvocationUnitTest.kt
@@ -36,7 +36,7 @@ import java.lang.ClassCastException
 
 private val javaLangString = libraryClassPool.getClass("java/lang/String")
 private val javaLangStringBuilder = libraryClassPool.getClass("java/lang/StringBuilder")
-private val valueFactory = ParticularValueFactory(DetailedArrayValueFactory(), ParticularReferenceValueFactory())
+private val valueFactory = ParticularValueFactory(DetailedArrayValueFactory(ParticularReferenceValueFactory()), ParticularReferenceValueFactory())
 private val stringExecutor = StringReflectionExecutor(libraryClassPool)
 private val invocationUnit = ExecutingInvocationUnit.Builder().build(valueFactory, libraryClassPool)
 private fun Int.toValue(): Value =